fix collisions problems. add mosquito. new player bullet pattern. many fixes/enhancements

+++ core/src/fr/poulpogaz/jam/stage/EnemyAction.java
@@ -2,15 +2,15 @@ package fr.poulpogaz.jam.stage;
 
 public final class EnemyAction<P> {
 
-    private final float start;
+    private final int start;
     private final P pattern;
 
-    public EnemyAction(float start, P pattern) {
+    public EnemyAction(int start, P pattern) {
         this.start = start;
         this.pattern = pattern;
     }
 
-    public float start() {
+    public int start() {
         return start;
     }
